FIX: 뱃지 수정

+++ backend/src/main/java/com/backend/db/entity/Badge.java
@@ -19,7 +19,7 @@ public class Badge {
 
     @Id
     @Column(name = "badge_sequence")
-    private Long badgeSequence;
+    private Integer badgeSequence;
     private String description;
     private String path;
